Update script path(s) in the P21Play database.
If the play database is using the same script path as the production
database, there is a chance that form Datastreams printed from the
production database will be appended to a Datastream file (.cfd) if it
already exists. This is due to the Datastream filenames being based off
of a `counter` value (`id` = `SCRPT`), which could potentially
"intersect" beween the two databases.  Write _all_ Datastreams from the
P21Play databse to a separate folder to avoid this.